idcp: Reimplement reservation wrappers
The IDCP code contains two wrappers for reservations; one for outgoing
reservations and one for ingoing reservations. Besides duplicating code,
and having inconsistent names, ingoing reservations don't have access to
notifcations via OSCARSNotifyOnlyImpl. This prevents correctly
implementing the IDCP protocol for ingoing reservations.

The existing implementation didn't follow the IDCP 1.1 specification
and consequently failed to work with Oscars 0.6. In particular the
implementation didn't use the correct state names and it didn't
trigger the notification events at the right point in time.

This patch more or less reimplements the wrappers. Common code is
factored out into the IdcpReservation base class, while code
specific to ingoing or outgoing reservations is in two subclasses.
Both hook into the OSCARSNotifyOnlyImpl to receive notifications
from the IDC peer.

The new code also provides the correct reservation creation time
on notifications.

The reply to reservation modification requests is fixed (the
reservation details are mandatory according to the spec).

The patch also fixes a memory leak: reservations were never removed
from the static embedded map of the wrapper.

The wrappers do not drive the autobahn internal reservation state
machine. Thus failures to setup the path in the IDC peer does
cause the Autobahn reservation to be aborted. This is no different
from the previous code. Fixing this is outside the scope of this
patch.

Testing done: Tested against a single Oscars 0.6 IDC peer. Created
a reservation in Oscars and verified that it is created, activated
and torn down in Autobahn. The opposite scenario Autobahn to Oscars
was tested as well. Cancellation was tested, as well as reservation
modification for reservations created in Oscars.

I am not convinced all failure cases are coverred, but doing so
likely requires hooking into the autobahn state machine anyway. Thus
I consider this outside the scope of this patch.

Unit tests pass.