Fix race condition in ServiceScheduler
ServiceScheduler was subject to the following race condition:

- run() would check whether the priorityCancel, cancel and exec queues are empty. Assume they are empty.
- Before run() enters the synchronized section an element is added to either queue.
- run() enters the synchronized block. 

The notification will be lost as run() was not waiting at the time the element was inserted in the queue. The consequence is that the scheduler will not schedule the request until another element is inserted into one of the queues.

The fix is to place the condition check inside the synchronization, thus making it part of the critical section.